Compliance: refactor handleServices duplicate
