+++ src/Components/CreateImageWizard/steps/Oscap/components/PolicySelector.tsx
@@ -18,7 +18,6 @@ import {
 import { useAppDispatch, useAppSelector } from '../../../../../store/hooks';
 import {
   Filesystem,
-  Services,
   useGetOscapCustomizationsForPolicyQuery,
   useLazyGetOscapCustomizationsForPolicyQuery,
 } from '../../../../../store/imageBuilderApi';
@@ -26,13 +25,9 @@ import {
   addKernelArg,
   addPartition,
   changeCompliance,
-  changeDisabledServices,
-  changeEnabledServices,
   changeFileSystemConfigurationType,
-  changeMaskedServices,
   clearKernelAppend,
   clearPartitions,
-  removePackage,
   selectCompliancePolicyID,
   selectCompliancePolicyTitle,
   selectDistribution,
@@ -92,7 +87,8 @@ const PolicySelector = () => {
   const hasWslTargetOnly = useHasSpecificTargetOnly('wsl');
   const dispatch = useAppDispatch();
   const [isOpen, setIsOpen] = useState(false);
-  const { clearCompliancePackages, handlePackages } = useSelectorHandlers();
+  const { clearCompliancePackages, handlePackages, handleServices } =
+    useSelectorHandlers();
 
   const {
     data: policies,
@@ -171,12 +167,6 @@ const PolicySelector = () => {
         dispatch(addPartition(partition));
       }
     }
-  };
-
-  const handleServices = (services: Services | undefined) => {
-    dispatch(changeEnabledServices(services?.enabled || []));
-    dispatch(changeMaskedServices(services?.masked || []));
-    dispatch(changeDisabledServices(services?.disabled || []));
   };
 
   const handleKernelAppend = (kernelAppend: string | undefined) => {
